+++ BizHawk.Emulation.Cores/Consoles/Nintendo/N64/N64.IMemoryDomains.cs
@@ -25,49 +25,49 @@ namespace BizHawk.Emulation.Cores.Nintendo.N64
 
 			IntPtr memPtr = api.get_memory_ptr(id);
 
-			Func<int, byte> peekByte;
-			Action<int, byte> pokeByte;
+			Func<long, byte> peekByte;
+			Action<long, byte> pokeByte;
 
 			if (swizzled)
 			{
-				peekByte = delegate(int addr)
+				peekByte = delegate(long addr)
 				{
 					if (addr < 0 || addr >= size)
 					{
 						throw new ArgumentOutOfRangeException();
 					}
 
-					return Marshal.ReadByte(memPtr, (addr ^ 3));
+					return Marshal.ReadByte(memPtr, (int)(addr ^ 3));
 				};
-				pokeByte = delegate(int addr, byte val)
+				pokeByte = delegate(long addr, byte val)
 				{
 					if (addr < 0 || addr >= size)
 					{
 						throw new ArgumentOutOfRangeException();
 					}
 
-					Marshal.WriteByte(memPtr, (addr ^ 3), val);
+					Marshal.WriteByte(memPtr, (int)(addr ^ 3), val);
 				};
 			}
 			else
 			{
-				peekByte = delegate(int addr)
+				peekByte = delegate(long addr)
 				{
 					if (addr < 0 || addr >= size)
 					{
 						throw new ArgumentOutOfRangeException();
 					}
 
-					return Marshal.ReadByte(memPtr, (addr));
+					return Marshal.ReadByte(memPtr, (int)(addr));
 				};
-				pokeByte = delegate(int addr, byte val)
+				pokeByte = delegate(long addr, byte val)
 				{
 					if (addr < 0 || addr >= size)
 					{
 						throw new ArgumentOutOfRangeException();
 					}
 
-					Marshal.WriteByte(memPtr, (addr), val);
+					Marshal.WriteByte(memPtr, (int)(addr), val);
 				};
 			}
 
@@ -116,15 +116,15 @@ namespace BizHawk.Emulation.Cores.Nintendo.N64
 			}
 
 
-			Func<int, byte> peekByte;
-			Action<int, byte> pokeByte;
+			Func<long, byte> peekByte;
+			Action<long, byte> pokeByte;
 
-			peekByte = delegate(int addr)
+			peekByte = delegate(long addr)
 			{
 				return api.m64p_read_memory_8((uint)addr);
 			};
 
-			pokeByte = delegate(int addr, byte val)
+			pokeByte = delegate(long addr, byte val)
 			{
 				api.m64p_write_memory_8((uint)addr, val);
 			};
@@ -132,7 +132,7 @@ namespace BizHawk.Emulation.Cores.Nintendo.N64
 			_memoryDomains.Add(new MemoryDomain
 				(
 					name: "System Bus",
-					size: 0, // special case for full 32bit memorydomain
+					size: 0, // special case for full 32bit memorydomain // int to long TODO: remove this special case!
 					endian: MemoryDomain.Endian.Big,
 					peekByte: peekByte,
 					pokeByte: pokeByte
